+++ .eslintrc.js
@@ -1,21 +1,30 @@
+'use strict';
+
 module.exports = {
-  "extends": ["ash-nazg/sauron-node"],
-  "parserOptions": {
-    "sourceType": "module"
+  extends: ['ash-nazg/sauron-node'],
+  parserOptions: {
+    sourceType: 'module'
   },
-  "plugins": [],
-  "env": {
-    "node": false,
-    "browser": true
+  plugins: [],
+  env: {
+    node: false,
+    browser: true
   },
   settings: {
     polyfills: [
-      "Object.assign",
-      "Promise.reject",
-      "Promise.resolve"
+      'Object.assign',
+      'Promise.reject',
+      'Promise.resolve'
     ]
   },
-  "overrides": [
+  overrides: [
+    {
+      files: '.eslintrc.js',
+      extends: ['plugin:node/recommended-script'],
+      rules: {
+        'import/no-commonjs': 0
+      }
+    },
     {
       files: '*.html',
       rules: {
@@ -33,19 +42,21 @@ module.exports = {
         headers: true,
         credentials: true,
         statusCallback: true,
-        retrievalCallback: true,
+        retrievalCallback: true
       },
       rules: {
         'import/unambiguous': 0,
         'jsdoc/require-jsdoc': 0,
-        'no-unused-vars': ['error', {varsIgnorePattern: 'json|status|retrieval'}]
+        'no-unused-vars': ['error', {
+          varsIgnorePattern: 'json|status|retrieval'
+        }]
       }
     },
     {
-      "files": ["test/index.js"],
+      files: ['test/index.js'],
       globals: {
-        assert: "readonly",
-        postJSON: "readonly"
+        assert: 'readonly',
+        postJSON: 'readonly'
       },
       env: {
         mocha: true
@@ -55,7 +66,7 @@ module.exports = {
       }
     }
   ],
-  "rules": {
-    "indent": ["error", 2, {"outerIIFEBody": 0}]
+  rules: {
+    indent: ['error', 2, {outerIIFEBody: 0}]
   }
 };