+++ src/index.js
@@ -108,6 +108,7 @@ function postJSON (url, bodyObject, cb, errBack) {
     dataObject.body = JSON.stringify(bodyObject);
   }
   dataObject.credentials = credentials;
+  // istanbul ignore next
   let ret = (typeof fetch !== 'undefined' ? fetch : postJSON.fetch)(
     url, dataObject
   ).then(statusCb).then(retrievalCb);