+++ CHANGES.md
@@ -1,5 +1,18 @@
 # CHANGES for postJSON
 
+## ?
+
+- Build: Use "json" extension for RC
+- Linting (ESLint): As per latest ash-nazg / ESLint 7
+- Testing: Check non-object signature, no credentials, object signature
+   with separate body, check callbacks and errBack, missing body, bad
+   server path, non-200 error code; brings coverage to 100%
+- Testing: Add nyc for coverage and use chai/register-assert
+- npm: Indicate in CLI for the browser test the URL that is to be opened
+- npm: Update `rollup-plugin-babel` to `@rollup/plugin-babel`
+    and make explicit `babelHelpers` value of `bundled`
+- npm: Update devDeps. and peerDeps. (core-js-bundle)
+
 ## 0.6.0
 
 - Breaking change: Change core-js-bundle to a peerDep.